+++ src/lib.rs
@@ -514,6 +514,7 @@ pub mod modular_fdis;
 pub mod non_zeros_grid;
 pub mod pass_group;
 pub mod pass_group_hf;
+pub mod per_channel_non_zeros;
 pub mod predictors;
 pub mod toc;
 pub mod vardct;